requirements.txt
customtkinter==6.5.1
tk==0.1.0
openai==1.0.0
xmldiff==2.1.0
openpyxl==3.1.2
python-docx==0.8.11
reportlab==4.0.0
Pillow==10.0.0
